requirements.txt
# Основные зависимости
aiogram>=3.2.0
python-dotenv>=1.0.0
loguru>=0.7.2

# База данных
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# HTTP и парсинг
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
feedparser>=6.0.10

# Дополнительные инструменты
python-dateutil>=2.8.2
pytz>=2023.3
aiofiles>=23.2.1

# Инструменты для работы с данными
pandas>=2.1.4
numpy>=1.26.2

# Инструменты для тестирования
pytest>=7.4.3
pytest-asyncio>=0.23.2
